+++ sync/_helpers.sh
@@ -1,0 +1,16 @@
+#!/usr/bin/env bash
+
+set -e
+set -u
+set -o pipefail
+
+# root dir of the git repository
+REPO_DIR=$(git rev-parse --show-toplevel) ; readonly repo_dir
+
+SCRIPT_DIR_REL=".${SCRIPT_DIR#"${REPO_DIR}"}" ; readonly SCRIPT_DIR
+
+# root dir of the generated helm chart
+CHART_DIR="${REPO_DIR}/helm/proxmox-csi-plugin" ; readonly CHART_DIR
+
+# root dir of the vendir synced chart
+VENDIR_SYNC_DIR="${REPO_DIR}/vendor/proxmox-csi-plugin/charts/proxmox-csi-plugin" ; readonly VENDIR_SYNC_DIR