+++ sync/patches/helpers/patch.sh
@@ -4,14 +4,16 @@ set -o errexit
 set -o nounset
 set -o pipefail
 
-repo_dir=$(git rev-parse --show-toplevel) ; readonly repo_dir
-script_dir=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd ) ; readonly script_dir
+SCRIPT_DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd ) ; readonly SCRIPT_DIR
+source "${SCRIPT_DIR}"/../../_helpers.sh
 
-cd "${repo_dir}"
+echo "Updating _helpers.tpl"
 
-readonly script_dir_rel=".${script_dir#"${repo_dir}"}"
+cd "${REPO_DIR}"
+
+readonly SCRIPT_DIR_REL=".${SCRIPT_DIR#"${REPO_DIR}"}"
 
 set -x
-git apply "${script_dir_rel}/_templates___helpers.tpl.patch"
+git apply "${SCRIPT_DIR_REL}/_templates___helpers.tpl.patch"
 
 { set +x; } 2>/dev/null